requirements.txt
eventregistry==9.1
fastapi==0.115.8
favicon==0.7.0
langchain-community==0.3.16
langchain-core==0.3.33
langchain-cohere==0.4.2
mysql-connector-python==9.2.0
pandas==2.2.3
pymongo==4.11
python-dotenv==1.0.1
schedule==1.2.2
tiktoken==0.8.0
uvicorn==0.34.0
